+++ sample_app/app/models/user.rb
@@ -1,3 +1,0 @@
-class User < ApplicationRecord
-  has_one_attached :avatar
-end