+++ sample_app/config/initializers/session_store.rb
@@ -1,1 +1,0 @@
-Rails.application.config.session_store :dynamodb_store, key: '_sample_app_session'