+++ sample_app_old/config/initializers/session_store.rb
@@ -1,0 +1,2 @@
+options = { key: '_sample_app_session' }
+Rails.application.config.session_store :dynamo_db_store, **options